% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look_busy.R
\name{look_busy}
\alias{look_busy}
\title{Look like you're working hard when you're hardly working}
\usage{
look_busy(
  minutes = Inf,
  speed = 1,
  background_job = FALSE,
  end = invisible(NULL)
)
}
\arguments{
\item{minutes}{How long to keep this up for.
Defaults to \code{Inf}, which runs forever.}

\item{speed}{How fast to produce output.
The default value, \code{1},
takes between 0.1 and 0.5 seconds to produce a calculation
and between 1 and 5 seconds to produce a status update.}

\item{background_job}{Whether to produce output in in a background job
(if \code{TRUE}) or the console (if \code{FALSE}).
Defaults to \code{FALSE}.
Running in a background process requires the \link[job:job]{job} package.}

\item{end}{Code to run after \code{minutes}.
If \code{minutes} is \code{Inf}, \code{end} is never run.}
}
\value{
\code{end} if \code{minutes} is finite.
}
\description{
Look like you're working hard when you're hardly working
}
\examples{
look_busy()
}
