% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play-package.R
\docType{package}
\name{play-package}
\alias{play}
\alias{play-package}
\title{play: Play games in the console}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Includes a number of games that can be played in the R console.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rossellhayes/play}
  \item Report bugs at \url{https://github.com/rossellhayes/play/issues}
}

}
\author{
\strong{Maintainer}: Alexander Rossell Hayes \email{alexander@rossellhayes.com} (\href{https://orcid.org/0000-0001-9412-0457}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Alan Beale \email{12dicts@pobox.com} (Word lists) [data contributor]
  \item Kevin Atkinson \email{kevina@users.sourceforge.net} (Word lists) [data contributor]
  \item Luis Von Ahn (Word lists) [data contributor]
  \item Flavia Rossell Hayes [illustrator]
  \item Kristin Bott \email{kristin.bott@posit.co} (look_busy() status messages) [contributor]
  \item Daniel Chen \email{daniel.chen@posit.co} (\href{https://orcid.org/0000-0003-3857-1741}{ORCID}) (look_busy() status messages) [contributor]
  \item Steven Smallberg \email{steven.smallberg@posit.co} (look_busy() status messages) [contributor]
}

}
\keyword{internal}
