% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play}
\alias{play}
\title{Play a game in the console}
\usage{
play(game = NULL, ...)
}
\arguments{
\item{game}{The name of a game to play.
If \code{NULL}, you can choose a game interactively in the console.}

\item{...}{Additional arguments passed to the chosen game.}
}
\value{
Calls the selected game function.
}
\description{
Interact with the game by typing commands into the console and
pressing enter.
}
\examples{
play()

play("poker")
play("hangman")
play("jumble")
play("2048")

play("poker", colors = 4)
play("hangman", difficulty = "easy")
play("jumble", difficulty = "hard")
play("2048", size = 5)
}
