% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_flip.R
\name{play_coin_flip}
\alias{play_coin_flip}
\title{Flip a coin}
\usage{
play_coin_flip(n = NULL, animate = rlang::is_interactive())
}
\arguments{
\item{n}{How many coins to flip.
Defaults to \code{NULL}, which interactively asks how many to flip.}

\item{animate}{If \code{TRUE}, play an animation before revealing the result.
Defaults to \code{TRUE} if the session is \link{interactive} and \code{FALSE} otherwise.}
}
\description{
Flip a coin
}
\examples{
play_coin_flip(1)

if (rlang::is_interactive()) play_coin_flip()
}
