% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shut_the_box.R
\name{play_shut_the_box}
\alias{play_shut_the_box}
\alias{shut_the_box}
\title{Play a game of shut the box in the console}
\usage{
play_shut_the_box()
}
\value{
Generates interactive output in the console.
}
\description{
Your goal is to flip down all nine tiles numbered from 1 to 9.
Each turn, you roll two dice and flip down any number of tiles that add up
to the same number as the sum of your dice.
After you have flipped down the 7, 8, and 9 tiles, you may choose whether to
roll one or two dice.
You win the game if you flip down all nine tiles!
But if you ever can't flip down a combination of tiles that sums to your
die roll, it's game over.
}
\examples{
play_shut_the_box()
}
