% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelling_bee.R
\name{play_spelling_bee}
\alias{play_spelling_bee}
\alias{spelling_bee}
\title{Play a game of Spelling Bee in the console}
\source{
The default word list is derived from
\href{http://wordlist.aspell.net/12dicts/}{12dicts} created by
\href{http://www.wyrdplay.org/}{Alan Beale}.

This word list is passed through the
\href{https://www.cs.cmu.edu/~biglou/resources/}{Offensive/Profane Word List}
created by \href{https://www.cs.cmu.edu/~biglou/}{Luis Von Ahn} to filter out
potentially unwanted words.
}
\usage{
play_spelling_bee()
}
\value{
Generates interactive output in the console.
}
\description{
Inspired by the \href{https://www.nytimes.com/puzzles/spelling-bee}{New York Times Spelling Bee},
created by Sam Ezersky.
}
\section{Problematic words?}{
You can report any words you find objectionable to
\url{https://github.com/rossellhayes/hangman/issues}.

Please report:
\itemize{
\item offensive words
\item words that deal with uncomfortable topics
\item proper nouns
\item text that is not a common English word
}
}

\examples{
play_spelling_bee()
}
