% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play}
\alias{play}
\title{Play a game in the console}
\usage{
play(game = NULL, ...)
}
\arguments{
\item{game}{The name of a game to play.
If \code{NULL}, you can choose a game interactively in the console.}

\item{...}{Additional arguments passed to the chosen game.}
}
\value{
Calls the selected game function.
}
\description{
Interact with the game by typing commands into the console and
pressing enter.
}
\examples{
if (rlang::is_interactive()) play()

play("magic_8_ball")
}
